+++ CHANGES.rst
@@ -20,6 +20,9 @@
 1.12.0 (unreleased)
 ===================
 
+- Fixed ``tpf.plot()`` and ``tpf.interact_sky()`` to reflect that Kepler and
+  TESS pixel coordinates refer to pixel centers. [#755]
+
 - Fixed a bug in ``tpf.interact()`` which caused the pixel selection to be off
   by half a pixel. The bug was introduced in v1.11.0. [#754]